% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire.R
\docType{data}
\name{fire}
\alias{fire}
\title{Fire Damage Dataset}
\format{
A data frame with 15 rows and 2 columns:
\describe{
  \item{DISTANCE}{Numeric. The distance from the reference point (e.g., fire station) in unspecified units.}
  \item{DAMAGE}{Numeric. The amount of damage caused by fire, in unspecified units.}
}
}
\source{
Simulated or collected data (source unknown).
}
\usage{
fire
}
\description{
A dataset containing information on fire damage in relation to the distance from a certain reference point (e.g., fire station or water source).
}
\keyword{datasets}
