% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{Plot a Normal Distribution Curve with Shaded Area}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{Mean of the normal distribution.}

\item{sigma}{Standard deviation of the normal distribution.}

\item{a}{Cutoff point where the probability is calculated.}
}
\value{
A list containing the mean, standard deviation, cutoff, and computed probability.
}
\description{
This function plots a normal distribution curve with a given mean and standard deviation,
shades the area under the curve from -∞ to a given value, and calculates the probability P(X ≤ a).
}
\details{
utils::globalVariables(c("x"))  # Prevents NOTE about x being undefined
}
\examples{
myncurve(mu = 0, sigma = 1, a = 1.5)
}
