% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myplot.R
\name{myplot}
\alias{myplot}
\title{Piecewise Regression for Spruce Data}
\usage{
myplot(data, breakpoint = 18)
}
\arguments{
\item{data}{A dataframe containing the variables "Height" and "BHDiameter".}

\item{breakpoint}{The breakpoint for the piecewise function (default is 18).}
}
\value{
A list containing the linear model summary and the plot.
}
\description{
This function fits a piecewise linear regression model to the spruce dataset
and plots the results.
}
