% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myntickets.R
\name{ntickets}
\alias{ntickets}
\title{Calculate the Optimal Number of Tickets to Sell}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{Integer. The total number of available seats on the flight.}

\item{gamma}{Numeric. The allowed probability of overbooking (i.e., more passengers show up than available seats).}

\item{p}{Numeric. The probability that an individual passenger shows up.}
}
\value{
A named list containing:
\describe{
  \item{n_disc}{Optimal number of tickets to sell using the discrete distribution.}
  \item{n_cont}{Optimal number of tickets to sell using the normal approximation.}
  \item{N}{Input value for total available seats.}
  \item{p}{Input probability of a passenger showing up.}
  \item{gamma}{Input probability of overbooking.}
}
}
\description{
This function determines the optimal number of tickets to sell for a flight
with `N` available seats, given the probability of a passenger showing up (`p`)
and the probability of overbooking (`gamma`). The function uses both a
discrete probability distribution and a normal approximation.
}
\examples{
ntickets(N = 400, gamma = 0.02, p = 0.95)

}
